FXML files for the home and menu options
Hi guys! 
So what I did was the home screen that will come up when a user enters his/her employee number and password. The Home screen has two buttons one that will lead to the MenuOptions screen and the second button will lead to another screen that will show the abilities a manager/owner has.  The Second FXML file is the MenuOptions this one has the three buttons one for breakfast, one for lunch and the last one for dinner. I will continue to work on more screens and connecting them. We can talk on Friday about more screens and their options. 
Best, 
Ximena