FXML Files
Hi guys!

So I added the FXML files that I did for the past week. All of the breakfast and mexican food stuff is completed. I can work on the Lunch so it can be ALL done. After this is a matter of connecting it to the database and making sure everything has prices and it shows a completed order.